Move common NURBS function into utils.nurbs_common module.
